fix: CI build fails because parent file resolution is wrong
