+++ package.json
@@ -12,7 +12,7 @@
     "tsc:watch": "tsc -w",
     "prepublishOnly": "tsc -d",
     "test": "tsc && mocha --timeout 10000 ./lib/**/__tests__/**.js",
-    "test:debug": "tsc && mocha --debug ./lib/**/__tests__/**.js",
+    "test:debug": "tsc && mocha --inspect ./lib/**/__tests__/**.js",
     "print": "npm run tsc && node ./lib/print.js",
     "print:sample1": "npm run tsc && node ./lib/print.js ./src/__tests__/data/ColumnHigherOrderComponent.tsx simple",
     "lint": "eslint -c lint.json 'src/**/*.{ts,tsx}'",
@@ -34,12 +34,12 @@
     "@types/react": "^16.9.34",
     "@types/source-map-support": "^0.4.1",
     "chai": "^4.1.2",
-    "eslint": "^7.5.0",
+    "eslint": "^7.32.0",
     "husky": "^0.14.3",
     "install": "^0.13.0",
     "lint-staged": "^7.3.0",
     "lodash": "^4.17.15",
-    "mocha": "^5.2.0",
+    "mocha": "^9.1.2",
     "prettier": "^1.19.1",
     "prop-types": "^15.6.2",
     "react": "^16.4.2",
